Modificado delete de extendedUser

+++ backend/src/main/java/es/project/service/impl/ExtendedUserServiceImpl.java
@@ -1,6 +1,7 @@
 package es.project.service.impl;
 
 import es.project.domain.ExtendedUser;
+import es.project.domain.User;
 import es.project.repository.ExtendedUserRepository;
 import es.project.repository.UserRepository;
 import es.project.service.ExtendedUserService;
@@ -91,6 +92,11 @@ public class ExtendedUserServiceImpl implements ExtendedUserService {
     @Override
     public void delete(Long id) {
         log.debug("Request to delete ExtendedUser : {}", id);
+        Optional<ExtendedUser> extendedUserOptional = extendedUserRepository.findById(id);
+        if (!extendedUserOptional.isPresent()) {
+            return;
+        }
         extendedUserRepository.deleteById(id);
+        userRepository.deleteById(id);
     }
 }
